rabbitmq-components.mk: Compare branch names before updating `rabbitmq-components.mk`
To avoid a copy of `rabbitmq-components.mk` from rabbitmq-common's
`v3.6.x` branch to another component's `master` branch, we want to
compare branch names of both repositories.

This situation may occur when checking out the `v3.6.x` in the Umbrella
using `gmake up BRANCH=v3.6.x` but for instance some plugins don't have
such a branch. In this case, they remain on the previous branch. We
don't want to overwrite their `rabbitmq-components.mk` if this happens.

[#154761483]

+++ deps/rabbit_common/mk/rabbitmq-components.mk
@@ -343,7 +343,16 @@ rabbitmq-components-mk:
 	@:
 else
 rabbitmq-components-mk:
+ifeq ($(FORCE),yes)
 	$(gen_verbose) cp -a $(UPSTREAM_RMQ_COMPONENTS_MK) .
+else
+	$(gen_verbose) if test -d .git && test -d $(DEPS_DIR)/rabbit_common/.git; then \
+		upstream_branch=$$(LANG=C git -C $(DEPS_DIR)/rabbit_common branch --list | awk '/^\* \(.*detached / {ref=$$0; sub(/.*detached [^ ]+ /, "", ref); sub(/\)$$/, "", ref); print ref; exit;} /^\* / {ref=$$0; sub(/^\* /, "", ref); print ref; exit}'); \
+		local_branch=$$(LANG=C git branch --list | awk '/^\* \(.*detached / {ref=$$0; sub(/.*detached [^ ]+ /, "", ref); sub(/\)$$/, "", ref); print ref; exit;} /^\* / {ref=$$0; sub(/^\* /, "", ref); print ref; exit}'); \
+		test "$$local_branch" = "$$upstream_branch" || exit 0; \
+	fi; \
+	cp -a $(UPSTREAM_RMQ_COMPONENTS_MK) .
+endif
 ifeq ($(DO_COMMIT),yes)
 	$(verbose) git diff --quiet rabbitmq-components.mk \
 	|| git commit -m 'Update rabbitmq-components.mk' rabbitmq-components.mk
